requirements.txt
flask
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-analytics-data
pandas
matplotlib
seaborn
scikit-learn